refactor: noticedetail 해당 폴더 이동
