+++ src/components/notice/ui/noticeDetailModal/CustomModal/CustomModal.styled.ts
