+++ src/components/notice/ui/noticeDetailModal/CustomModal/CustomModal.tsx
