--- src/pages/noticedetail/utils/usePriceUtils.ts
@@ -1,3 +1,0 @@
-// export const calculatePriceChange = (originalPay: number, currentPay: number): number => {
-//   return ((currentPay - originalPay) / originalPay) * 100;
-// };
