--- src/pages/noticedetail/utils/useRecentPosts.ts
@@ -1,1 +1,0 @@
-// 추후 로직 분리
