+++ src/widgets/noticedetail/index.ts
@@ -1,0 +1,3 @@
+export { DetailWidget } from './ui/DetailWidget/DetailWidget';
+export { RecentPostsWidget } from './ui/RecentPostsWidget/RecentPostsWidget';
+export { ShopDetailWidget } from './ui/ShopDetailWidget/ShopDetailWidget';
