+++ src/widgets/noticedetail/ui/DetailWidget/DetailWidget.styled.ts
