+++ src/widgets/noticedetail/ui/DetailWidget/DetailWidget.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import * as S from './DetailWidget.styled';
-import { NoticeData } from '@/shared/types/post'; 
+import { NoticeData } from '@/shared/types/post';
 
 interface DetailWidgetProps {
   noticeData?: NoticeData;
