+++ src/widgets/noticedetail/ui/RecentPostsWidget/RecentPostWidget.tsx
