+++ src/widgets/noticedetail/ui/RecentPostsWidget/RecentPostsWidget.styled.ts
