+++ src/widgets/noticedetail/ui/ShopDetailWidget/ShopDetailWidget.styled.ts
