+++ src/widgets/noticedetail/ui/ShopDetailWidget/ShopDetailWidget.tsx
@@ -5,10 +5,10 @@ import PostPrice from '@/shared/components/Post/PostPrice/PostPrice';
 import PostInform from '@/shared/components/Post/PostInform/PostInform';
 import formatWorkTime from '@/shared/utils/formatWorkTime';
 import CustomButton from '@/shared/components/Button/CustomButton/CustomButton';
-import CustomModal from '../../component/CustomModal/CustomModal';
+import CustomModal from '../../../../components/notice/ui/noticeDetailModal/CustomModal/CustomModal';
 import checkImg from '@/assets/check.svg';
 import cautionImg from '@/assets/caution.svg';
-import { useHandleModal } from '../../utils/useHandleModal';
+import { useHandleModal } from '../../../../shared/utils/useHandleModal';
 import { NoticeData } from '@/shared/types/post';
 import { useModal } from '@/shared/store/useModal';
 import { useRouter } from 'next/router';
